feat: 룰렛 시스템 가중치 소수점 지원 및 12슬롯 최적화
- weight 컬럼을 INT에서 DECIMAL(10,2)로 변경하여 정밀한 확률 설정 지원
- 룰렛 슬롯을 정확히 12개로 표준화하고 각 슬롯별 확률 최적화 (총 100%)
- 기본 아이템 목록을 12개로 확장하고 균형잡힌 확률 분포 적용 (1%~20%)
- 룰렛 애니메이션 종료 시 정확한 위치에 당첨 아이템이 배치되도록 개선
- 아이템 목록 표시 시 가중치와 확률을

+++ src/main/kotlin/com/lukehemmin/lukeVanilla/System/Roulette/RouletteData.kt
@@ -48,7 +48,7 @@ data class RouletteItem(
     val itemDisplayName: String?,
     val itemAmount: Int,
     val itemData: String?, // JSON 형태
-    val weight: Int,
+    val weight: Double, // 가중치 (소수점 2자리까지 지원)
     val enabled: Boolean,
     val createdAt: Timestamp,
     val updatedAt: Timestamp
